feat: add ERC20 and native token permission schemas
- Introduced new schemas for ERC20 token allowances, periodic permissions, and streaming permissions.
- Added native token schemas for allowances, periodic permissions, and streaming permissions.
- Created utility sections for handling justification, permission info, and review summaries.
- Implemented token approval revocation schema to manage various token approval methods.
- Added handling for unknown permission types with appropriate schema structure.
- Refactored utility functions for parsing hex permission amounts.
