+++ packages/7715-permission-types/src/index.ts
@@ -25,4 +25,49 @@ export {
   type DeployedContractsByName,
   type PermissionDecoderConfig,
 } from './permissions';
-export * from './permissions/schema';
+export {
+  DAY,
+  FORTNIGHT,
+  HOUR,
+  MAX_UINT256,
+  MONTH,
+  SECOND,
+  WEEK,
+  YEAR,
+  ALL_METAMASK_FACILITATOR_ADDRESSES,
+  METAMASK_FACILITATOR_ADDRESSES,
+  METAMASK_FACILITATOR_ADDRESSES_DEV,
+  areOnlyMetaMaskFacilitatorAddresses,
+  isMetaMaskFacilitatorAddress,
+  convertAmountPerSecondToAmountPerPeriod,
+  convertMillisecondsToSeconds,
+  formatPermissionPeriodDuration,
+  getPeriodFrequencyValueTranslationKey,
+  parseHexPermissionAmount,
+  getPermissionSchemaEntry,
+} from './permissions/schema';
+export type {
+  AccountField,
+  AddressField,
+  AmountField,
+  DateField,
+  DividerElement,
+  ExpiryField,
+  FieldView,
+  I18nFunction,
+  JustificationField,
+  ListField,
+  NetworkField,
+  OriginField,
+  PermissionRenderContext,
+  PermissionSchemaEntry,
+  PermissionSchemaRegistry,
+  RawTextField,
+  ReviewFieldView,
+  RuleAddressField,
+  SchemaElement,
+  SchemaSection,
+  TextField,
+  TokenResolution,
+  TokenVariant,
+} from './permissions/schema';
