+++ packages/7715-permission-types/src/permissions/schema/nativeTokenAllowance.ts
@@ -1,0 +1,20 @@
+import {
+  allowanceDetailsSection,
+  justificationSection,
+  permissionInfoSection,
+  requireStartTime,
+  reviewSummaryAccountSection,
+} from './sections';
+import type { PermissionSchemaEntry } from './types';
+
+export const nativeTokenAllowanceSchema: PermissionSchemaEntry = {
+  tokenVariant: 'native',
+  tokenResolution: { kind: 'native' },
+  validate: requireStartTime,
+  sections: [
+    justificationSection,
+    permissionInfoSection,
+    allowanceDetailsSection('native-token-allowance-details-section'),
+    reviewSummaryAccountSection,
+  ],
+};
